Quordle now favors later indexes
This IS exploiting the fact that words.txt is sorted with "guesses on top, actual answers (but also valid guesses) on bottom"

But what's happening right now is that it's explicitly AVOIDING actual answers

There's an easy to solution to this, and it's to sort the array. Writing a comparator would be easy, but... meh.