The second test is wrong.
First: We know the Python tests are correct. If THOSE are wrong, we're 100% screwed. VERY very screwed.

However, Test 2 is only backed by the previous versions. It's possible those versions were also bugged, so using this five-nested for loop we can run ApplyGuess and activate Debug Mode to compare what it outputs to System.out with what the Debugger says.